game-basics Wave A: FruitCamera + ExplodyFruitModifier + gameplay statics
First batch of the game-basics port (static, RE-driven). Ports unimplemented
v1.6.1 binary logic that the // TODO markers specified; unblocks SuperFruitControl.

- FruitCamera::UpdateCamera (0x1edf24): m_field14c cast dropped (v1.6.1 writes
  none), m_LookAt <- engine Zero Vec3 global, SinTransition zoom-k + roll
  constants resolved from DAT, and the unconditional m_bDirty=true per-frame
  write restored (binary strb [.,#0x108]; was omitted -> stale camera matrices).
- ExplodyFruitModifier::FruitSplosion (0x135620): FruitWasKilled subscribed to
  Fruit::m_OnKilled; file-static chain-combo list (s_pChainHead, binary .bss
  0x2D9FE8) with MakeCombo+AddToCurrentScore on count>2; burst Update/DrawOrder;
  ADingoAteMyBaby.
- MissControl::MakeEmAllDissappear (0x19dd74): fade-alpha clamp to exact
  DAT_0019ddd4=0x3d8da741; SuperFruitControl::PushBombsAway radial bomb push.

asm-verify v1.6.1: 215 cosmetic / 993 deferred / 41 defunct / 0 divergence.
Two FIX-NEEDED fidelity bugs (camera dirty-flag, clamp constant bit-pattern)
caught by the strict triage pass and fixed before commit.

+++ src/game/ExplodyFruitModifier.h
@@ -61,7 +61,7 @@ public:
     public:
         // Own fields (HUDControl3d base = 0x7c)
         Mortar::Entity* m_pEntity;  // +0x7c
-        uint32_t        m_Const80;  // +0x80: DAT_135868
+        float           m_Const80;  // +0x80: DAT_135868=0; timer accumulator (dt*wavedt)
         float           m_p0;       // +0x84: forceMin
         float           m_p1;       // +0x88: forceInc
         float           m_p2;       // +0x8c: forceMax
@@ -78,14 +78,14 @@ public:
         ~FruitSplosion() override;
 
         void Update(float dt) override;
-        // TODO: 0x00135620 — FruitSplosion::Draw / DrawOrder body not yet ported
+        // RE-ported: 0x00134E00 — thunks to HUDControl3d::DrawOrder (no custom body).
         void DrawOrder(const Vec3& hudScale, int layerMask) override;
 
-        // Delegate target: called when sliced fruit is killed
-        // TODO: 0x00135620 — FruitWasKilled body not yet ported
+        // Delegate target: called when tracked fruit is killed (nulls m_pEntity).
+        // RE-ported: 0x00134DEC
         void FruitWasKilled(Fruit* fruit);
-        // Delegate target: removal callback
-        // TODO: 0x00135620 — ADingoAteMyBaby body not yet ported
+        // Delegate target: removal callback registered on chain head.
+        // RE-ported: 0x00134E04
         void ADingoAteMyBaby(HUDControl* ctrl);
     };
